Passing activities layers; more docs; more tests
