+++ examples/src/main/scala/com/example/heartbeatingactivity/HeartbeatingActivityBatchWorker.scala
@@ -5,7 +5,7 @@ import zio.temporal._
 import zio.temporal.worker._
 import zio.temporal.workflow._
 import zio.logging.backend.SLF4J
-import zio.temporal.activity.ZActivityRunOptions
+import zio.temporal.activity.{ZActivityImplementationObject, ZActivityRunOptions}
 
 object HeartbeatingActivityBatchWorker extends ZIOAppDefault {
   val TaskQueue = "HeartbeatingActivityBatch"
@@ -14,8 +14,18 @@ object HeartbeatingActivityBatchWorker extends ZIOAppDefault {
     Runtime.removeDefaultLoggers ++ SLF4J.slf4j
 
   override def run: ZIO[ZIOAppArgs with Scope, Any, Any] = {
+    val activitiesLayer: URLayer[
+      RecordLoader with RecordProcessor with ZActivityRunOptions[Any],
+      List[ZActivityImplementationObject[_]]
+    ] = ZLayer.collectAll(
+      List(
+        ZActivityImplementationObject.layer(RecordProcessorActivityImpl.make),
+        ZActivityImplementationObject.layer(ReporterActivityImpl.make)
+      )
+    )
+
     val registerWorkflows = ZWorkerFactory.newWorker(TaskQueue) @@
-      ZWorker.addActivityImplementationService[RecordProcessorActivity] @@
+      ZWorker.addActivityImplementationsLayer(activitiesLayer) @@
       ZWorker.addWorkflow[HeartbeatingActivityBatchWorkflowImpl].fromClass
 
     val program = for {
