+++ integration-tests/src/test/scala/zio/temporal/WorkflowSpec.scala
@@ -2,15 +2,11 @@ package zio.temporal
 
 import zio._
 import zio.logging.backend.SLF4J
-import zio.temporal.activity.{ZActivityImplementationObject, ZActivityOptions}
 import zio.temporal.fixture._
 import zio.temporal.testkit._
 import zio.temporal.worker._
 import zio.temporal.workflow._
 import zio.test._
-
-import java.util.concurrent.atomic.{AtomicBoolean, AtomicInteger, AtomicReference}
-import scala.collection.mutable.ListBuffer
 
 object WorkflowSpec extends BaseTemporalSpec {
   override val bootstrap: ZLayer[Any, Any, TestEnvironment] =
@@ -384,7 +380,9 @@ object WorkflowSpec extends BaseTemporalSpec {
         result <- ZWorkflowStub.execute(
                     workflow.doSomething(0)
                   )
-      } yield assertTrue(result == "Done")
+      } yield {
+        assertTrue(result == "Done")
+      }
 
     }.provideTestWorkflowEnv,
     test("run workflow with ZIO untyped activity") {
@@ -461,326 +459,6 @@ object WorkflowSpec extends BaseTemporalSpec {
         )
       }
 
-    }.provideTestWorkflowEnv @@ TestAspect.flaky,
-    test("run workflow with successful sagas") {
-      ZTestWorkflowEnvironment.activityRunOptionsWithZIO[Any] { implicit activityRunOptions =>
-        val taskQueue      = "saga-queue"
-        val successFunc    = (_: String, _: BigDecimal) => ZIO.succeed(Done())
-        val expectedResult = BigDecimal(5.0)
-
-        for {
-          workflowId <- ZIO.randomWith(_.nextUUID)
-
-          _ <- ZTestWorkflowEnvironment.newWorker(taskQueue) @@
-                 ZWorker.addWorkflow[SagaWorkflowImpl].fromClass @@
-                 ZWorker.addActivityImplementation(new TransferActivityImpl(successFunc, successFunc))
-          _ <- ZTestWorkflowEnvironment.setup()
-          sagaWorkflow <- ZTestWorkflowEnvironment.newWorkflowStub[SagaWorkflow](
-                            ZWorkflowOptions
-                              .withWorkflowId(workflowId.toString)
-                              .withTaskQueue(taskQueue)
-                              .withWorkflowRunTimeout(10.seconds)
-                          )
-          result <- ZWorkflowStub.executeWithTimeout(5.seconds)(
-                      sagaWorkflow.transfer(TransferCommand("from", "to", expectedResult))
-                    )
-        } yield {
-          assertTrue(result == expectedResult)
-        }
-      }
-
-    }.provideTestWorkflowEnv,
-    test("run workflow with saga compensations") {
-      ZTestWorkflowEnvironment.activityRunOptionsWithZIO[Any] { implicit activityRunOptions =>
-        val taskQueue = "saga-queue"
-
-        val From = "from"
-        val To   = "to"
-
-        val compensated = new AtomicBoolean(false)
-        val withdrawn   = new AtomicInteger(0)
-        val error       = TransferError("fraud")
-        val withdrawFunc = (_: String, _: BigDecimal) =>
-          ZIO.succeed {
-            withdrawn.incrementAndGet()
-            Done()
-          }
-        val depositFunc: (String, BigDecimal) => IO[TransferError, Done] = {
-          case (From, _) =>
-            ZIO.succeed {
-              compensated.set(true)
-              Done()
-            }
-          case _ =>
-            ZIO.fail(error)
-        }
-
-        val amount = BigDecimal(5.0)
-
-        for {
-          workflowId <- ZIO.randomWith(_.nextUUID)
-
-          _ <- ZTestWorkflowEnvironment.newWorker(taskQueue) @@
-                 ZWorker
-                   .addWorkflow[SagaWorkflowImpl]
-                   .withOptions(
-                     // provide more activity options here
-                     ZWorkflowImplementationOptions.default.withActivityOptions(
-                       "Deposit" ->
-                         ZActivityOptions
-                           .withStartToCloseTimeout(5.seconds)
-                           .withScheduleToCloseTimeout(60.seconds)
-                           .withRetryOptions(
-                             ZRetryOptions.default
-                               .withMaximumAttempts(1)
-                               .withDoNotRetry(nameOf[TransferError])
-                           )
-                     )
-                   )
-                   .fromClass @@
-                 ZWorker.addActivityImplementation(new TransferActivityImpl(depositFunc, withdrawFunc))
-
-          _ <- ZTestWorkflowEnvironment.setup()
-          sagaWorkflow <- ZTestWorkflowEnvironment.newWorkflowStub[SagaWorkflow](
-                            ZWorkflowOptions
-                              .withWorkflowId(workflowId.toString)
-                              .withTaskQueue(taskQueue)
-                              .withWorkflowRunTimeout(10.seconds)
-                          )
-          result <- ZWorkflowStub
-                      .execute(sagaWorkflow.transfer(TransferCommand(From, To, amount)))
-                      .either
-        } yield {
-          assert(result)(Assertion.isLeft) &&
-          assertTrue(
-            compensated.get(),
-            // activity options were applied correctly
-            withdrawn.get() == 1
-          )
-        }
-      }
-
-    }.provideTestWorkflowEnv,
-    test("run workflow with zasync") {
-      val taskQueue = "zasync-queue"
-
-      val order = new AtomicReference(ListBuffer.empty[(String, Int)])
-      val fooFunc = (x: Int) => {
-        println(s"foo($x)")
-        order.get += ("foo" -> x)
-        x
-      }
-      val barFunc = (x: Int) => {
-        println(s"bar($x)")
-        order.get += ("bar" -> x)
-        x
-      }
-
-      val x = 2
-      val y = 3
-
-      def runTest(n: Int) = for {
-        promiseWorkflow <- ZTestWorkflowEnvironment.newWorkflowStub[PromiseWorkflow](
-                             ZWorkflowOptions
-                               .withWorkflowId(s"zasync-spec-$n")
-                               .withTaskQueue(taskQueue)
-                               .withWorkflowRunTimeout(10.seconds)
-                           )
-        result <- ZWorkflowStub.execute(promiseWorkflow.fooBar(x, y))
-      } yield {
-        assertTrue(result == x + y)
-      }
-
-      for {
-        _ <- ZTestWorkflowEnvironment.newWorker(taskQueue) @@
-               ZWorker.addWorkflow[PromiseWorkflowImpl].fromClass @@
-               ZWorker.addActivityImplementation(new PromiseActivityImpl(fooFunc, barFunc))
-
-        _ <- ZTestWorkflowEnvironment.setup()
-
-        assertions <- ZIO
-                        .foreach(List.range(1, 21))(runTest)
-                        .map { results =>
-                          val actualResult = order.get.toList
-
-                          println(actualResult.toString)
-
-                          results.reduce(_ && _) && assertTrue(
-                            actualResult.size == 40,
-                            actualResult != List.fill(20)(List("foo" -> x, "bar" -> y))
-                          )
-                        }
-      } yield assertions
-    }.provideTestWorkflowEnv @@ TestAspect.flaky(3),
-    test("run parameterized workflow") {
-      val taskQueue = "parameterized-queue"
-
-      def runWorkflow[Input <: ParameterizedWorkflowInput](
-        stub:  ZWorkflowStub.Of[ParameterizedWorkflow[Input]]
-      )(input: Input
-      ): TemporalIO[List[ParameterizedWorkflowOutput]] = {
-        ZIO.logInfo("Executing parameterized workflow") *>
-          ZWorkflowStub.execute(stub.parentTask(input))
-      }
-
-      for {
-        uuid <- ZIO.randomWith(_.nextUUID)
-        _ <- ZTestWorkflowEnvironment.newWorker(taskQueue) @@
-               ZWorker.addWorkflow[SodaWorkflow].from(new SodaWorkflowImpl) @@
-               ZWorker.addWorkflow[SodaChildWorkflow].from(new SodaChildWorkflowImpl) @@
-               ZWorker.addWorkflow[JuiceWorkflow].from(new JuiceWorkflowImpl) @@
-               ZWorker.addWorkflow[JuiceChildWorkflow].from(new JuiceChildChildWorkflowImpl)
-
-        _ <- ZTestWorkflowEnvironment.setup()
-
-        sodaWf <- ZTestWorkflowEnvironment.newWorkflowStub[SodaWorkflow](
-                    ZWorkflowOptions
-                      .withWorkflowId(s"soda/$uuid")
-                      .withTaskQueue(taskQueue)
-                  )
-
-        juiceWf <- ZTestWorkflowEnvironment.newWorkflowStub[JuiceWorkflow](
-                     ZWorkflowOptions
-                       .withWorkflowId(s"juice/$uuid")
-                       .withTaskQueue(taskQueue)
-                   )
-
-        sodaResult  <- runWorkflow(sodaWf)(ParameterizedWorkflowInput.Soda("coke"))
-        juiceResult <- runWorkflow(juiceWf)(ParameterizedWorkflowInput.Juice("orange"))
-
-      } yield {
-        assertTrue(
-          sodaResult.sortBy(_.message) == List(
-            ParameterizedWorkflowOutput("Providing with soda: coke"),
-            ParameterizedWorkflowOutput("Providing with soda: coke"),
-            ParameterizedWorkflowOutput("Providing with soda: coke")
-          ),
-          juiceResult.sortBy(_.message) == List(
-            ParameterizedWorkflowOutput("Providing with orange juice (1L)"),
-            ParameterizedWorkflowOutput("Providing with orange juice (2L)"),
-            ParameterizedWorkflowOutput("Providing with orange juice (3L)")
-          )
-        )
-      }
-    }.provideTestWorkflowEnv,
-    test("Runs workflows with timers") {
-      val taskQueue = "payment-queue"
-      for {
-        workflowId <- ZIO.randomWith(_.nextUUID)
-        _ <- ZTestWorkflowEnvironment.newWorker(taskQueue) @@
-               ZWorker.addWorkflow[PaymentWorkflowImpl].fromClass
-
-        // Setup the test environment
-        _ <- ZTestWorkflowEnvironment.setup()
-
-        // Create a workflow stub
-        paymentWorkflow <- ZTestWorkflowEnvironment.newWorkflowStub[PaymentWorkflow](
-                             ZWorkflowOptions
-                               .withWorkflowId(workflowId.toString)
-                               .withTaskQueue(taskQueue)
-                               // Set workflow timeout
-                               .withWorkflowRunTimeout(20.minutes)
-                           )
-
-        // Start the workflow stub
-        _ <- ZWorkflowStub.start(
-               paymentWorkflow.processPayment(42.1)
-             )
-        // Skip time
-        _ <- ZTestWorkflowEnvironment.sleep(10.minutes + 30.seconds)
-        // Get the workflow result
-        isFinished <- paymentWorkflow.result[Boolean]
-      } yield assertTrue(!isFinished)
-    }.provideTestWorkflowEnv,
-    test("Workflow retry") {
-      val taskQueue = "retry-queue"
-      for {
-        workflowId <- ZIO.randomWith(_.nextUUID)
-        _ <- ZTestWorkflowEnvironment.newWorker(taskQueue) @@
-               ZWorker.addWorkflow[RetryWorkflowImpl].fromClass
-
-        _ <- ZTestWorkflowEnvironment.setup()
-
-        retryWorkflow <- ZTestWorkflowEnvironment.newWorkflowStub[RetryWorkflow](
-                           ZWorkflowOptions
-                             .withWorkflowId(workflowId.toString)
-                             .withTaskQueue(taskQueue)
-                             .withWorkflowRunTimeout(10.second)
-                         )
-
-        result <- ZWorkflowStub.execute(retryWorkflow.withRetry(2)).either
-      } yield assert(result)(Assertion.isRight(Assertion.equalTo("OK")))
-    }.provideTestWorkflowEnv,
-    test("Workflow memo") {
-      val taskQueue = "memo-queue"
-      for {
-        workflowId <- ZIO.randomWith(_.nextUUID)
-        _ <- ZTestWorkflowEnvironment.newWorker(taskQueue) @@
-               ZWorker.addWorkflow[MemoWorkflowImpl].fromClass
-
-        _ <- ZTestWorkflowEnvironment.setup()
-
-        memoWorkflow <- ZTestWorkflowEnvironment.newWorkflowStub[MemoWorkflow](
-                          ZWorkflowOptions
-                            .withWorkflowId(workflowId.toString)
-                            .withTaskQueue(taskQueue)
-                            .withMemo("zio" -> "temporal")
-                            .withWorkflowRunTimeout(10.second)
-                        )
-
-        result <- ZWorkflowStub.execute(memoWorkflow.withMemo("zio"))
-      } yield assert(result)(Assertion.isSome(Assertion.equalTo("temporal")))
-    }.provideTestWorkflowEnv,
-    test("Register workflows using addWorkflowImplementations") {
-      val taskQueue = "child-2-queue"
-
-      for {
-        workflowId <- ZIO.randomWith(_.nextUUID)
-        _ <- ZTestWorkflowEnvironment.newWorker(taskQueue) @@
-               ZWorker.addWorkflowImplementations(
-                 ZWorkflowImplementationClass[GreetingWorkflowImpl],
-                 ZWorkflowImplementationClass[GreetingChildImpl]
-               )
-        _ <- ZTestWorkflowEnvironment.setup()
-        greetingWorkflow <- ZTestWorkflowEnvironment.newWorkflowStub[GreetingWorkflow](
-                              ZWorkflowOptions
-                                .withWorkflowId(workflowId.toString)
-                                .withTaskQueue(taskQueue)
-                                .withWorkflowRunTimeout(10.second)
-                            )
-        result <- ZWorkflowStub.execute(
-                    greetingWorkflow.getGreeting("Vitalii")
-                  )
-      } yield assertTrue(result == "Hello Vitalii!")
-    }.provideTestWorkflowEnv,
-    test("Register activities using addActivityImplementations") {
-      val taskQueue = "multi-activities-workflow-queue"
-
-      ZTestWorkflowEnvironment.activityRunOptionsWithZIO[Any] { implicit activityRunOptions =>
-        for {
-          workflowId <- ZIO.randomWith(_.nextUUID)
-          _ <- ZTestWorkflowEnvironment.newWorker(taskQueue) @@
-                 ZWorker.addWorkflowImplementations(
-                   List(
-                     ZWorkflowImplementationClass[MultiActivitiesWorkflowImpl]
-                   )
-                 ) @@
-                 ZWorker.addActivityImplementations(
-                   ZActivityImplementationObject(new ZioActivityImpl()),
-                   ZActivityImplementationObject(ComplexTypesActivityImpl())
-                 )
-          _ <- ZTestWorkflowEnvironment.setup()
-          multiWorkflow <- ZTestWorkflowEnvironment.newWorkflowStub[MultiActivitiesWorkflow](
-                             ZWorkflowOptions
-                               .withWorkflowId(workflowId.toString)
-                               .withTaskQueue(taskQueue)
-                               .withWorkflowRunTimeout(10.second)
-                           )
-          result <- ZWorkflowStub.execute(
-                      multiWorkflow.doSomething("Vitalii")
-                    )
-        } yield assertTrue(result == "Echoed Vitalii, list=2")
-      }
-    }.provideTestWorkflowEnv
+    }.provideTestWorkflowEnv @@ TestAspect.flaky
   )
 }
