+++ integration-tests/src/test/scala/zio/temporal/fixture/MultiActivitiesWorkflow.scala
@@ -48,3 +48,69 @@ class MultiActivitiesWorkflowImpl extends MultiActivitiesWorkflow {
     echoed + s", list=${list.size}"
   }
 }
+
+// Activity with dependencies
+object ReporterService {
+  val make: ULayer[ReporterService] =
+    ZLayer.succeed(new ReporterService)
+}
+class ReporterService {
+  def notify(message: String): Task[Unit] =
+    ZIO.logInfo(s"Reporting message: $message")
+}
+
+@activityInterface
+trait ActivityWithDependencies {
+  def reportSomething(n: Int): Unit
+}
+
+object ActivityWithDependenciesImpl {
+  val make: URLayer[ReporterService with ZActivityRunOptions[Any], ActivityWithDependencies] =
+    ZLayer.fromFunction(ActivityWithDependenciesImpl(_: ReporterService)(_: ZActivityRunOptions[Any]))
+}
+
+case class ActivityWithDependenciesImpl(
+  reporterService:  ReporterService
+)(implicit options: ZActivityRunOptions[Any])
+    extends ActivityWithDependencies {
+  override def reportSomething(n: Int): Unit =
+    ZActivity.run {
+      reporterService.notify(s"given a number=$n")
+    }
+}
+
+// More workflows
+class EvenMoreMultiActivitiesWorkflowImpl extends MultiActivitiesWorkflow {
+  private val logger = ZWorkflow.makeLogger
+
+  private val complexTypesActivity = ZWorkflow.newActivityStub[ComplexTypesActivity](
+    ZActivityOptions.withStartToCloseTimeout(5.seconds)
+  )
+
+  private val activityWithDependencies = ZWorkflow.newActivityStub[ActivityWithDependencies](
+    ZActivityOptions.withStartToCloseTimeout(5.seconds)
+  )
+
+  private val zioActivity = ZWorkflow.newActivityStub[ZioActivity](
+    ZActivityOptions.withStartToCloseTimeout(5.seconds)
+  )
+
+  override def doSomething(input: String): String = {
+    logger.info("Invoking first activity...")
+
+    val echoed = ZActivityStub.execute(
+      zioActivity.echo(input)
+    )
+
+    logger.info("Invoking second activity...")
+    val list = ZActivityStub.execute(
+      complexTypesActivity.complexList
+    )
+
+    ZActivityStub.execute(
+      activityWithDependencies.reportSomething(list.size)
+    )
+
+    echoed + s", list=${list.size}"
+  }
+}
